Cập nhật bảo mật

+++ admin/ajax/ajax_xoa_de_tai.php
@@ -1,36 +1,32 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: Ly Thanh Ngo
- * Date: 16/04/2018
- * Time: 7:32 AM
- */
-?>
-<?php
 include_once("../../config.php");
 session_start();
 if (isset($_SESSION['tdn']) && isset($_SESSION['pas'])) {
     $ketnoi = new clsKetnoi();
     if (!($ketnoi->checklogin($_SESSION['tdn'],$_SESSION['pas']))) {
-        echo '<script type="text/javascript">location.href = "'.$qlkh['HOSTADMIN'].'"</script>';
+        trangchu($qlkh['HOSTADMIN']);
         exit();
     }
 }else{
-    echo '<script type="text/javascript">location.href = "'.$qlkh['HOSTADMIN'].'"</script>';
+    trangchu($qlkh['HOSTADMIN']);
 }
 $result = Array(
     'trangthai' => 0
 );
-function xoadetai($iddt){
+function xoadetai($iddt,$idnd){
     $ketnoi = new clsKetnoi();
     $conn = $ketnoi->ketnoi();
     $iddt = intval($iddt);
+    $idnd = intval($idnd);
     $q = "SELECT IDDT FROM detai WHERE IDDT = '$iddt' AND TRANGTHAI != 'Đã nghiệm thu'";
     $eq = mysqli_query($conn,$q);
     $rq = mysqli_num_rows($eq);
     if ($rq == 0) return false;
-    $sql = "DELETE FROM `detai` WHERE `IDDT` = '$iddt';";
-    $sql.= "DELETE FROM `loaihinhnghiencuu` WHERE `IDDT` = '$iddt';";
+    $sql = "DELETE FROM `detai` WHERE `IDDT` = '$iddt' AND EXISTS (SELECT IDND FROM thanhviendetai WHERE IDND = $idnd AND IDDT = $iddt AND TRACHNHIEM=N'Chủ nhiệm');";
+    $test = mysqli_query($conn,$sql);
+    $demdong = mysqli_affected_rows($conn);
+    if ($demdong==0) return false;
+    $sql= "DELETE FROM `loaihinhnghiencuu` WHERE `IDDT` = '$iddt';";
     $sql.= "DELETE FROM `linhvuckhoahoc` WHERE `IDDT` = '$iddt';";
     $sql.= "DELETE FROM `thanhviendetai` WHERE `IDDT` = '$iddt';";
     $sql.= "DELETE FROM `tochucthamgia` WHERE `IDDT` = '$iddt';";
@@ -44,7 +40,7 @@ function xoadetai($iddt){
     else
         return false;
 }
-if (xoadetai($_POST['id'])) {
+if (xoadetai($_POST['id'],$_SESSION['_idnd'])) {
     $result['trangthai'] = 1;
     echo json_encode($result);
     exit();
